+++ main.csv.go
@@ -18,10 +18,32 @@ func getAssetsFromCSV(csvFile, assetType string) (bool, []map[string]string) {
 	}
 	// automatically call Close() at the end of current method
 	defer file.Close()
-	//
-	r := csv.NewReader(file)
 
+    bom := make([]byte, 3)
+    file.Read(bom)
+    if bom[0] == 0xEF && bom[1] == 0xBB && bom[2] == 0xBF {
+        // BOM Detected, continue with feeding the file fmt.Println("BOM")
+    } else {
+        // No BOM Detected, reset the file feed
+        file.Seek(0,0)
+    }
+
+    //because the json configuration loader cannot handle runes, code here to convert string to rune-array and getting first item
+	r := csv.NewReader(file)
+	if CSVImportConf.CSVCommaCharacter != "" {
+        CSVCommaRunes := []rune(CSVImportConf.CSVCommaCharacter)
+        r.Comma = CSVCommaRunes[0]
+        //r.Comma = ';'
+    }
+    
+	if CSVImportConf.CSVLazyQuotes {
+        r.LazyQuotes = true
+    }
+	if CSVImportConf.CSVFieldsPerRecord > 0 {
+        r.FieldsPerRecord = CSVImportConf.CSVFieldsPerRecord
+    }
 	var header []string
+    
 	for {
 		record, err := r.Read()
 		if err == io.EOF {