+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.2 (September 20th, 2017)
+
+Features:
+
+  - Added code to deal with UTF-8 BOM at the beginning of some CSV files
+  - Allow for setting a different field separator (Comma), LazyQuotes and FieldsPerRecord within the import configuration
+
 ## 1.0.1 (August 3rd, 2017)
 
 Features: